pyproject.toml
[project]
name = "dataverse-mcp"
version = "0.1.0"
description = "A read-only MCP server for querying Microsoft Dataverse environments during development"
readme = "README.md"
requires-python = ">=3.10"
authors = [
    {name = "Ryan James"},
]
license = "MIT"
keywords = [
    "mcp",
    "model-context-protocol",
    "dataverse",
    "power-platform",
    "copilot",
    "llm",
]
classifiers = [
    "Development Status :: 5 - Production/Stable",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "azure-identity>=1.25.3",
    "mcp[cli]>=1.27.0",
    "powerplatform-dataverse-client>=0.1.0b7",
]

[project.urls]
Homepage = "https://github.com/ryanmichaeljames/dataverse-mcp"
Repository = "https://github.com/ryanmichaeljames/dataverse-mcp"
Issues = "https://github.com/ryanmichaeljames/dataverse-mcp/issues"
Changelog = "https://github.com/ryanmichaeljames/dataverse-mcp/blob/main/CHANGELOG.md"

[project.scripts]
dataverse-mcp = "dataverse_mcp.server:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[dependency-groups]
dev = [
    "pytest>=8.0",
    "build>=1.0",
    "twine>=5.0",
]
